backend/requirements.txt
Flask>=1.1.2
Flask-SQLAlchemy>=2.4.4
Flask-Migrate>=2.5.3
psycopg2>=2.8.5
pytest>=6.0.1
Flask-Mail>=0.9.1
APScheduler>=3.6.3
Flask-APScheduler>=1.11.0
pika>=1.1.0